rename component and add migration db
