+++ package.json
@@ -5,6 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
+    "backend": "npx json-server backends/db.json --port=4000",
     "build": "vite build",
     "lint": "eslint .",
     "preview": "vite preview"
@@ -12,10 +13,12 @@
   "dependencies": {
     "@ant-design/icons": "^5.5.1",
     "antd": "^5.21.6",
+    "json-server": "^1.0.0-beta.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.3.0",
     "react-infinite-scroll-component": "^6.1.0",
+    "react-redux": "^9.2.0",
     "react-router-dom": "^6.28.1"
   },
   "devDependencies": {
