+++ src/MainComponent/FriendStatusContentDetailsComponent.jsx
