+++ src/main.jsx
@@ -3,7 +3,6 @@ import { createRoot } from 'react-dom/client'
 import { App } from './App'
 import "./index.css";
 
-
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <App />
